\name{metafetcher-package}
\alias{metafetcher-package}
\alias{metafetcher}
\docType{package}
\title{
\packageTitle{metafetcher}
}
\description{
\packageDescription{metafetcher}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metafetcher}
\packageIndices{metafetcher}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{metafetcher}

Maintainer: \packageMaintainer{metafetcher}
}
\references{
aaa
}
\keyword{package}
\seealso{
\link[ROSETTA]{bioinf.icm.uu.se}
}
\examples{
  library(metafetcher)
}
