% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{resolve}
\alias{resolve}
\title{Resolves missing metabolome IDs and other attributes in dataframe}
\usage{
resolve(df.discovered)
}
\arguments{
\item{df.discovered}{A dataframe}
}
\value{
A list containing some statistics and the same dataframe, extended with missing data.
}
\description{
Resolves missing metabolome IDs and other attributes in dataframe
}
\examples{
df_from_csv <- read.csv("meta_ids.csv", stringsAsFactors=FALSE)
resp <- resolve(df_from_csv)
df.out <- resp$df
view(df.out)
}
