% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{resolve_single_id}
\alias{resolve_single_id}
\title{Resolves missing external IDs belonging to a single metabolome ID}
\usage{
resolve_single_id(start_db_tag, start_db_id)
}
\arguments{
\item{start_db_tag}{The type of the ID - which database it belongs to? All lowercase and must end in '_id'}

\item{start_db_id}{The metabolome ID in the given database}
}
\value{
A list containing some statistics and a dataframe, that contains a single row with the resolved data.
}
\description{
Resolves missing external IDs belonging to a single metabolome ID
}
\examples{
# For example to start the search from https://www.ebi.ac.uk/chebi/searchId.do?chebiId=CHEBI:8077
resp <- resolve_single_id("chebi_id", "8077")
df.out <- resp$df
view(df.out)
}
